Move ResolveError -> core/resolver/errors.rs
Move activation_error, describe_path -> errors.rs
